# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ProgressDB API
  version: "0.1.0"
  description: |
    Author semantics and signing flow:
      - Protected endpoints require a cryptographically-verified author identity, supplied
        via the headers `X-User-ID` and `X-User-Signature` (HMAC-SHA256).
      - Backend callers holding a backend API key may request an HMAC signature for a user
        by calling `POST /v1/_sign`. Clients attach the returned signature to subsequent
        requests as `X-User-Signature` with `X-User-ID`.
      - For user-scoped operations the server uses the verified author derived from the
        signature middleware (`X-User-ID` + `X-User-Signature`) as the canonical author.
      - An explicit `author` query parameter is accepted only for trusted callers (role
        `admin` or `backend`) and is rejected for regular/front-end callers to prevent
        impersonation.
servers:
  - url: /
paths:
...
  /v1/messages:
    parameters:
      - $ref: '#/components/parameters/XUserID'
      - $ref: '#/components/parameters/XUserSignature'
      - $ref: '#/components/parameters/IncludeDeleted'
    get:
      summary: List messages for a thread
      parameters:
        - in: query
          name: thread
          schema:
            type: string
          description: Thread ID to list
        - in: query
          name: limit
          schema:
            type: integer
          description: Optional limit of most recent messages
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  thread:
                    type: string
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
    post:
      summary: Create a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /v1/messages/{id}:
    parameters:
      - $ref: '#/components/parameters/XUserID'
      - $ref: '#/components/parameters/XUserSignature'
    get:
      summary: Get latest message version by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Latest message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
    put:
      summary: Edit (append new version) a message by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses:
        "200":
          description: Updated message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
    delete:
      summary: Soft-delete a message by ID (append tombstone)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No content
  /v1/messages/{id}/versions:
    parameters:
      - $ref: '#/components/parameters/XUserID'
      - $ref: '#/components/parameters/XUserSignature'
    get:
      summary: List all stored versions for a message ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Versions list
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  versions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
  /v1/messages/{id}/reactions:
    parameters:
      - $ref: '#/components/parameters/XUserID'
      - $ref: '#/components/parameters/XUserSignature'
    get:
      summary: List reactions on a message
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  reactions:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        reaction:
                          type: string
    post:
      summary: Add or update a reaction on a message
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  description: identity id (opaque to server)
                reaction:
                  type: string
              required: [id, reaction]
      responses:
        "200":
          description: Updated message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /v1/messages/{id}/reactions/{identity}:
    parameters:
      - $ref: '#/components/parameters/XUserID'
      - $ref: '#/components/parameters/XUserSignature'
    delete:
      summary: Remove a reaction for an identity
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: identity
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No content
...
components:
  schemas:
    Message:
      type: object
      properties:
        id:
          type: string
          description: Message ID (server-generated if missing)
        thread:
          type: string
          description: Thread ID
        author:
          type: string
        ts:
          type: integer
          format: int64
          description: Timestamp in nanoseconds (server-filled if missing)
        body:
          description: Freeform JSON payload
          nullable: true
        reply_to:
          type: string
          description: Message ID this message replies to
        deleted:
          type: boolean
          description: Soft-delete flag
        reactions:
          type: object
          additionalProperties:
            type: string
      example:
        id: msg-1700000000-1
        thread: t1
        author: u1
        ts: 1700000000000000000
        body:
          text: "hello"
          credit_card:
            _enc: gcm
            v: "BASE64_NONCE_CIPHERTEXT"
...
  securitySchemes:
    AdminApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Provide an admin or backend API key in the X-API-Key header or as Bearer in the Authorization header.
  parameters:
    XUserID:
      name: X-User-ID
      in: header
      required: false
      schema:
        type: string
      description: |
        Verified user id. When present this header is used alongside `X-User-Signature`
        for frontend-signed requests. Backend callers holding a backend/admin API key
        may also supply `X-User-ID` without a signature to assert an author.
    XUserSignature:
      name: X-User-Signature
      in: header
      required: false
      schema:
        type: string
      description: |
        HMAC-SHA256 signature of the `X-User-ID` computed by a backend signer.
        Required for frontend callers (who do not hold backend keys). Backend callers
        that possess a backend/admin API key may omit this header and instead set
        `X-User-ID`.
...
    IncludeDeleted:
      name: include_deleted
      in: query
      required: false
      schema:
        type: boolean
      description: When true, include messages marked deleted (tombstones) in list responses.